Refactor getClueIds null check in GetButton component
Moved the null check for getClueIds inside the condition that checks if the id exists in get_clues. This will prevent errors for users that do not have an associated get_clues string array. Previously, if a user did not have any get_clues, the system would error out due to trying to run an includes check on an undefined value.