Add functionality to update 'get_people' field
This commit introduces a change to the 'GetButton.tsx' and 'useUserFunctions.ts' files that allows the 'get_people' field to be updated in the database when specific conditions are met. This was necessary to ensure the correct values are updated when the 'GetButton' component is used with the 'Earring' model. The 'updateUserGetPeople' function was added to supabase functions to facilitate these updates.