"Refactor GetButton to fix undefined clues issue"
Updated the GetButton component to prevent concatenation with undefined 'get_clues'. Previously, the code always concatenated the current model ID with the 'get_clues' variable, but if this variable did not exist, it resulted in the creation of a string starting with undefined. The current change now includes a check for the existence of 'get_clues' before concatenation, thereby preventing the creation of inconsistent IDs.