Remove result limit and add date button
The limit condition was removed from the "useRankingFunctions" hook in order to display all the results instead of just the first 10. Also, a date button was added in the ranking page for debugging purposes, it console logs the current datetime when clicked. This button will later be utilized for date/time sorting feature.