Update SVG and config files and refactor code
Added a new SVG file to the images directory, updated the references to this SVG in the Header component, and made minor tweaks to the styling. Also, removed `optimizeFonts` config from the next.js config as it's deprecated. Refactored `_document.tsx` for better readability and performance by removing unnecessary code related to font loading. These changes improve the load time and enhance the overall user experience.