Add new functionality and adjust layout
This commit introduces new functionality to display additional information about the creators and resources used in the project. A separate info page has been created for this purpose, and relevant components are included to display creator and resource details. Also, some adjustments have been made to the layout and styling of existing pages to provide a more streamlined user experience. These changes enhance the overall functionality and aesthetics of the project.