Increase maximum clues from 7 to 8 and add a new clue "document"
This commit increases the maximum number of clues from 7 to 8 to accommodate for a new clue "document". This change impacts the clue acquisition progress indicated on the progress page and extends the condition for routing to the home page if the current clue id is greater than the maximum number of clues.

Additionally, a new clue "document" is added to the list of clues in the scanning page, its AR marker pattern