Add support for "Document Clipboard with Pen" 3D model
This commit incorporates the "Document Clipboard with Pen" 3D model into the the project. The relevant files (license.txt, scene.gltf) and related changes in CopyrightText.tsx and ArModel.tsx are included. Now, the AR feature of the application is capable of rendering this new model. The 3D model is provided by Kami Rapacz under the CC-BY-4.0 license which necessitates displaying proper attribution in the implementation.