Add "Tools & Services" to the TechStack and GitHub link
A new category named "Tools & Services" has been introduced in the TechStack component of our application that lists some of the tools and services used in the project such as GitHub, GitHub Copilot, Figma, and Notion. Likewise, a GitHub icon link is incorporated in the 'result' page, providing users easy access to the project on GitHub. This update aims to improve transparency about the tech stack and provide an easy navigation to our GitHub repository.